requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0

python-dotenv==1.0.0
email-validator==2.1.1

pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
